Added online vs offline space.
For any room inside of "The Online Area" region, navigation will only
work via clicking on links. Links are a type of door that can be used to
separate rooms (pages) on a computer website. Turning on the computer
will bring you online - By bringing you to the Login Screen.

A basic structure of a Login Screen and Homepage exist for stubs of the
online space. There is an office and a Bathroom for testing navigation
in the real world space.